+++ src/DataStreams/XmlStream.php
@@ -22,13 +22,14 @@ require_once __DIR__ . '/../Xml/XmlParser.php';
 require_once __DIR__ . '/../Xml/NodeStrategyCollationDictionary.php';
 require_once __DIR__ . '/DataStream.php';
 
+use Rhubarb\Crown\Exceptions\ImplementationException;
 use Rhubarb\Crown\Xml\NodeStrategyCollationDictionary;
 use Rhubarb\Crown\Xml\XmlParser;
 
 /**
  * Scans an XML document for individual nodes and pops them off through a stream.
  */
-class XmlStream extends DataStream
+class XmlStream extends RecordStream
 {
     /**
      * @var
@@ -53,7 +54,7 @@ class XmlStream extends DataStream
         $this->xmlFilePath = $xmlFilePath;
     }
 
-    private function Open()
+    private function open()
     {
         if ($this->xmlParser == null) {
             $this->xmlParser = new XmlParser($this->xmlFilePath);
@@ -65,7 +66,7 @@ class XmlStream extends DataStream
 
     public function readNextItem()
     {
-        $this->Open();
+        $this->open();
 
         $available = $this->xmlParser->parseOne();
 
@@ -74,5 +75,10 @@ class XmlStream extends DataStream
         }
 
         return false;
+    }
+
+    public function appendItem($item)
+    {
+        throw new ImplementationException();
     }
 }